When profiling memory events in memcheck, event counts larget than INT_MAX have been observed in practice. Make counters 64 bit entities.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@15516

+++ memcheck/mc_include.h
@@ -320,7 +320,7 @@ enum {
    MCPE_LAST
 };
 
-extern UInt MC_(event_ctr)[MCPE_LAST];
+extern ULong MC_(event_ctr)[MCPE_LAST];
 
 #  define PROF_EVENT(ev)                           \
    do { tl_assert((ev) >= 0 && (ev) < MCPE_LAST);  \
